feat: move calculating project graph hash to own function
