+++ lib/block.js
@@ -14,16 +14,13 @@ const
 	isPathInside = require('is-path-inside');
 
 const
-	objectHash = require('node-object-hash'),
-	hashFiles = require('hash-files');
-
-const
 	resolve = require('./resolve'),
 	Declaration = require('./declaration');
 
 const
 	{blockTypeList} = /** @type {{blockTypeList: Array<string>}} */ require('./validators'),
-	{createSerializableMap, createSerializableSet, jsonReviver} = require('./helpers');
+	{createSerializableMap, createSerializableSet, jsonReviver} = require('./helpers'),
+	{calculateProjectGraphHash} = require('./graph-hash');
 
 const
 	filesCache = Object.create(null),
@@ -151,42 +148,24 @@ class Block {
 	 * Returns block manifests file by the specified names
 	 *
 	 * @param {Array<string>=} [names]
-	 * @param {{locPrefix: ?string}=} [opts]
+	 * @param {{lockPrefix: ?string}=} [opts]
 	 * @returns {!Promise<!Map<string, !Block>>}
 	 */
 	static async getAll(names, opts = {}) {
-		const
-			projectCWD = await resolve.cwd,
-			sourceDirs = resolve.sourceDirs;
+		const projectCWD = await resolve.cwd;
 
 		const componentsLockFile = path.join(
 			projectCWD,
 			[].concat(opts.lockPrefix || [], 'components-lock.json').join('.')
 		);
 
-		const globOpts = {
-			ignore: [
-				path.join(resolve.entry(), '/**'),
-				...sourceDirs.map((dir) => path.join(dir, '**/tmp/**'))
-			]
-		};
-
 		console.log('Calculating a project graph hash');
 
-		const projectHash = objectHash().hash({
-			lockPrefix: opts.lockPrefix || '',
-
-			srcHash: hashFiles.sync({
-				files: sourceDirs
-					.flatMap((dir) => glob.sync(path.join(dir, `/**/@(${blockTypeList.join('|')})-*/index.js`), globOpts))
-					.sort()
-			}),
-
-			projectFiles: glob.sync(path.join(sourceDirs[0], `/**/@(${blockTypeList.join('|')}-*/*.@(js|ts|styl|ss|ess)`), globOpts)
-				.sort()
-				.map((src) => path.relative(projectCWD, src)),
-
-			objToHash: this.objToHash
+		const projectHash = await calculateProjectGraphHash({
+			projectCWD,
+			lockPrefix: opts.lockPrefix,
+			sourceDirs: resolve.sourceDirs,
+			entryDir: resolve.entry()
 		});
 
 		console.log(`The project graph hash has been calculated`);
